requirements/requirements-dev.txt
# Test tools
pytest
pytest-django
coverage

# Linting & fomatting
flake8
black

# Security
bandit
safety

# Code complexity
radon